docs(playground): add loading component when loading playgound dynamically
